Completely restructured print output processing.
Width, precision and header for various items to be printed are now defined as a dict in main. This is passed to each function that needs to print, and then used within that function. It's a bit elaborate, but hopefully makes it easy to adjust an item's printing width and have everthing adapt everywhere.